Extract tests to separate files for mutations checkout shipping/billing address, checkout create
