+++ saleor/graphql/checkout/tests/test_checkout_billing_address_update.py
@@ -129,3 +129,53 @@ def test_checkout_billing_address_update_by_id_without_street_address_2(
     assert checkout.billing_address.postal_code == billing_address["postalCode"]
     assert checkout.billing_address.country == billing_address["country"]
     assert checkout.billing_address.city == billing_address["city"].upper()
+
+
+def test_checkout_billing_address_update(
+    user_api_client, checkout_with_item, graphql_address_data
+):
+    checkout = checkout_with_item
+    assert checkout.shipping_address is None
+    previous_last_change = checkout.last_change
+
+    query = """
+    mutation checkoutBillingAddressUpdate(
+            $id: ID, $billingAddress: AddressInput!) {
+        checkoutBillingAddressUpdate(
+                id: $id, billingAddress: $billingAddress) {
+            checkout {
+                token,
+                id
+            },
+            errors {
+                field,
+                message
+            }
+        }
+    }
+    """
+    billing_address = graphql_address_data
+
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "billingAddress": billing_address,
+    }
+
+    response = user_api_client.post_graphql(query, variables)
+    content = get_graphql_content(response)
+    data = content["data"]["checkoutBillingAddressUpdate"]
+    assert not data["errors"]
+    checkout.refresh_from_db()
+    assert checkout.billing_address is not None
+    assert checkout.billing_address.first_name == billing_address["firstName"]
+    assert checkout.billing_address.last_name == billing_address["lastName"]
+    assert (
+        checkout.billing_address.street_address_1 == billing_address["streetAddress1"]
+    )
+    assert (
+        checkout.billing_address.street_address_2 == billing_address["streetAddress2"]
+    )
+    assert checkout.billing_address.postal_code == billing_address["postalCode"]
+    assert checkout.billing_address.country == billing_address["country"]
+    assert checkout.billing_address.city == billing_address["city"].upper()
+    assert checkout.last_change != previous_last_change
